refactor(web): split the alerts wizard's steps and state machine
The last two allowlisted files in the alerts directory. Same directory,
pure moves with their doc comments:

- steps.tsx (846 -> 576) keeps the seven step bodies.
  stepPrimitives.tsx takes the shared presentation (StepTitle,
  ResultLine, FinalPill, Summary, Mono, Composed, Field, CopyRow);
  stepShared.ts takes the JSX-free K, StepProps and reasonText, which
  is what keeps stepPrimitives exporting components only
  (react-refresh/only-export-components); destinationKinds.ts takes the
  KINDS tile table with its hint, title and icon maps.
- AlertsWizard.tsx (801 -> 438) keeps the component, its rail and its
  announcer. wizardState.ts takes the state machine: Step,
  TOTAL_STEPS, DEFAULT_APPRISE_URL, Draft, WizardState, the Action
  union, EMPTY_DRAFT, isDuplicate, canNext, newDraft, initialState and
  reducer. It imports AlertsWizardProps type-only, so the pair is not a
  runtime cycle.

The wizard's own test now imports initialState/isDuplicate/reducer from
wizardState; nothing else changes. Four react-refresh disable comments
travelled with the reducer helpers and are dropped: the rule does not
apply in a JSX-free module, and a stale directive fails lint.

Both entries leave the size allowlist.

Verified on the rebuilt dev stack: the wizard opens at step 2 of 7,
Next stays disabled until a kind is picked, picking ntfy enables it,
the reducer advances to Details, typing fills the composed-URL panel,
Next stays disabled while Topic is empty, and Back preserves the picked
kind. No console errors.
