+++ web/src/pages/Settings/alerts/steps.tsx
@@ -1,35 +1,37 @@
 import {
 	APPRISE_SERVICES_URL,
-	type DestinationKind,
 	FIELDS,
-	type FieldDef,
 	parseDiscordWebhook,
 } from "@web-shared/alerts/composers";
 import { eventLabel, parseCsv } from "@web-shared/alerts/events";
 import { generateTopic } from "@web-shared/ntfy";
-import type { TFunction } from "i18next";
-import { type Dispatch, Fragment, type ReactNode, useState } from "react";
+import { Fragment, useState } from "react";
 import {
 	Bell,
 	CheckCircle2,
 	CheckSquare,
-	DiscordLogo,
-	Link,
 	ListChecks,
-	type LucideIcon,
-	Mail,
 	Pencil,
 	PlugZap,
 	Send,
 	Smartphone,
-	TelegramLogo,
-	XCircle,
 } from "@/lib/icons";
-import type { AlertEventDef, AlertStatus } from "../../../api/types";
-import { CopyButton } from "../../../components/CopyButton";
+import type { AlertEventDef } from "../../../api/types";
 import { AlertEventPicker } from "../AlertEventPicker";
-import { type Action, isDuplicate, type WizardState } from "./AlertsWizard";
 import { DestinationList } from "./DestinationList";
+import { KIND_HINT, KIND_ICON, KIND_TITLE, KINDS } from "./destinationKinds";
+import {
+	Composed,
+	CopyRow,
+	Field,
+	FinalPill,
+	Mono,
+	ResultLine,
+	StepTitle,
+	Summary,
+} from "./stepPrimitives";
+import { K, reasonText, type StepProps } from "./stepShared";
+import { isDuplicate } from "./wizardState";
 
 // The seven step bodies of the alerts wizard: prove apprise-api answers, pick
 // what kind of destination this is, fill in the parts that kind needs, deliver
@@ -37,71 +39,6 @@ import { DestinationList } from "./DestinationList";
 // write. They are presentation only; every gate lives in AlertsWizard's
 // canNext, so a step can never let itself through, and only the last one asks
 // the wizard to persist anything.
-
-const K = "settings.alerts.wizard";
-
-export interface StepProps {
-	state: WizardState;
-	dispatch: Dispatch<Action>;
-	t: TFunction;
-}
-
-// StepTitle names the step. It is a plain heading: the step change is announced
-// by the wizard's own live region (AlertsWizard's StepAnnouncer), which is
-// mounted once for the whole run so a step change mutates its text rather than
-// inserting a new region, and screen readers reliably miss the latter. The icon
-// chip beside the title is the same accent chip the settings sections and
-// dashboard cards wear, so each step opens like a section of the app rather
-// than a form field.
-function StepTitle({
-	id,
-	icon: Icon,
-	children,
-}: {
-	id?: string;
-	icon: LucideIcon;
-	children: ReactNode;
-}) {
-	return (
-		<div className="flex items-center gap-3">
-			<span className="ui-wizard-icon" aria-hidden="true">
-				<Icon size={18} />
-			</span>
-			<h3 className="text-base font-semibold text-(--text-primary)" id={id}>
-				{children}
-			</h3>
-		</div>
-	);
-}
-
-// ResultLine is one sentence reporting how something the operator just asked
-// for turned out, in the theme's success or error tone with the matching
-// glyph. The glyph is decoration: the sentence carries the meaning, and the
-// element's text is the sentence alone.
-function ResultLine({
-	ok,
-	testId,
-	role = "status",
-	children,
-}: {
-	ok: boolean;
-	testId: string;
-	role?: "status" | "alert";
-	children: ReactNode;
-}) {
-	const Icon = ok ? CheckCircle2 : XCircle;
-	return (
-		<p
-			role={role}
-			data-testid={testId}
-			data-ok={ok ? "true" : "false"}
-			className={`flex items-start gap-1.5 text-sm ${ok ? "ui-text-success" : "ui-text-error"}`}
-		>
-			<Icon size={16} className="mt-0.5 shrink-0" aria-hidden="true" />
-			<span>{children}</span>
-		</p>
-	);
-}
 
 export function StepApprise({
 	state,
@@ -165,45 +102,6 @@ export function StepApprise({
 		</>
 	);
 }
-
-// Model Hotel has no companion app of its own, so the phone tile is ntfy's.
-// (composers.ts still recognises a Bellhop topic when it describes a stored
-// destination that Front Desk set up.)
-const KINDS = [
-	"ntfy",
-	"telegram",
-	"discord",
-	"email",
-	"other",
-] as const satisfies readonly DestinationKind[];
-
-const KIND_HINT: Record<(typeof KINDS)[number], string> = {
-	ntfy: "kindNtfyHint",
-	telegram: "kindTelegramHint",
-	discord: "kindDiscordHint",
-	email: "kindEmailHint",
-	other: "kindOtherHint",
-};
-
-// Three tiles are named after the service itself, so they reuse the shared
-// settings.alerts.kind.* labels. Two need more than the bare name to be picked
-// correctly: "ntfy" means nothing until it says it is the phone app, and
-// "Apprise URL" is the catch-all rather than a service.
-const KIND_TITLE: Partial<Record<(typeof KINDS)[number], string>> = {
-	ntfy: "kindNtfyTitle",
-	other: "kindOtherTitle",
-};
-
-// One glyph per tile, so the five options can be told apart at a glance
-// before the titles are read: the two services with a logo wear it.
-const KIND_ICON: Record<(typeof KINDS)[number], LucideIcon> = {
-	ntfy: Smartphone,
-	telegram: TelegramLogo,
-	discord: DiscordLogo,
-	email: Mail,
-	other: Link,
-};
-
 export function StepKind({
 	state,
 	dispatch,
@@ -279,7 +177,6 @@ export function StepKind({
 		</>
 	);
 }
-
 export function StepDetails({ state, dispatch, t }: StepProps) {
 	const kind = state.draft.kind;
 	if (kind === null) return null;
@@ -395,7 +292,6 @@ export function StepDetails({ state, dispatch, t }: StepProps) {
 		</>
 	);
 }
-
 export function StepTest({
 	state,
 	t,
@@ -431,7 +327,6 @@ export function StepTest({
 		</>
 	);
 }
-
 export function StepDestinations({
 	state,
 	dispatch,
@@ -499,7 +394,6 @@ export function StepDestinations({
 		</>
 	);
 }
-
 export function StepEvents({
 	state,
 	dispatch,
@@ -564,7 +458,6 @@ export function StepEvents({
 		</>
 	);
 }
-
 export function StepFinish({
 	state,
 	t,
@@ -679,168 +572,5 @@ export function StepFinish({
 				</ResultLine>
 			)}
 		</>
-	);
-}
-
-// FinalPill reports the probe taken straight after the write. The card's own
-// status line covers a fourth state (nothing configured at all) that cannot
-// happen here: the wizard has just configured it.
-function FinalPill({ status, t }: { status: AlertStatus; t: TFunction }) {
-	const [variant, label] = !status.reachable
-		? (["ui-badge-error", "unreachable"] as const)
-		: !status.healthy
-			? (["ui-badge-warning", "issues"] as const)
-			: (["ui-badge-success", "reachable"] as const);
-	return (
-		<span
-			className={`ui-badge ${variant}`}
-			data-testid="wiz-done-pill"
-			title={status.detail}
-		>
-			{t(`settings.alerts.status.${label}`)}
-		</span>
-	);
-}
-
-// One labelled tile of the closing summary: what is about to be written, in
-// the operator's own words rather than as the settings keys it becomes.
-function Summary({
-	label,
-	testId,
-	children,
-}: {
-	label: string;
-	testId?: string;
-	children: ReactNode;
-}) {
-	return (
-		<div className="ui-detail-tile space-y-1.5 px-3 py-2.5">
-			<span className="block text-[11px] font-medium uppercase tracking-wider text-(--text-tertiary)">
-				{label}
-			</span>
-			<span className="block min-w-0" data-testid={testId}>
-				{children}
-			</span>
-		</div>
-	);
-}
-
-// Mono is an address in the theme's own mono face (Tailwind's font-mono is a
-// fixed stack and would ignore the Terminal style's JetBrains Mono). It may
-// break anywhere: an address has no word boundaries worth keeping.
-function Mono({ block, children }: { block?: boolean; children: ReactNode }) {
-	return (
-		<code
-			className={`${block ? "block " : ""}text-xs text-(--text-primary) select-all break-all`}
-			style={{ fontFamily: "var(--font-mono)" }}
-		>
-			{children}
-		</code>
-	);
-}
-
-// reasonText renders a server reason code. Codes the catalog does not cover
-// (and a failure that carried none) fall back to the caller's own wording
-// rather than leaking a raw key, or a sentence about the wrong thing, into the
-// dialog: a probe that could not be made is not a test that failed to deliver.
-function reasonText(code: string, t: TFunction, fallback: string): string {
-	return t(`settings.alerts.reason.${code}`, { defaultValue: fallback });
-}
-
-// The composed Apprise URL, shown from the moment the fields make a valid one.
-// It is the thing that gets tested and stored, so the operator sees it rather
-// than having to trust that the fields were assembled the way they expect.
-function Composed({ url, t }: { url: string; t: TFunction }) {
-	if (url === "") return null;
-	return (
-		<div className="ui-detail-tile space-y-1.5 px-3 py-2.5">
-			<span className="block text-[11px] font-medium uppercase tracking-wider text-(--text-tertiary)">
-				{t(`${K}.composedLabel`)}
-			</span>
-			{/* The theme's own mono face: Tailwind's font-mono is a fixed stack and
-			    would ignore the Terminal style's JetBrains Mono. */}
-			<code
-				className="block text-xs text-(--text-primary) select-all break-all"
-				data-testid="wiz-composed"
-				style={{ fontFamily: "var(--font-mono)" }}
-			>
-				{url}
-			</code>
-		</div>
-	);
-}
-
-function Field({
-	def,
-	label,
-	value,
-	onChange,
-	children,
-}: {
-	def: FieldDef;
-	label: string;
-	value: string;
-	onChange: (v: string) => void;
-	children?: ReactNode;
-}) {
-	const id = `wiz-field-${def.key}`;
-	return (
-		<div className="space-y-1.5">
-			<label
-				className="text-sm font-medium text-(--text-secondary)"
-				htmlFor={id}
-			>
-				{label}
-			</label>
-			<div className="flex items-center gap-2">
-				<input
-					id={id}
-					data-testid={id}
-					className="ui-input text-sm w-full"
-					// Secrets are hidden while they are typed, which is the only moment
-					// someone could be looking over a shoulder; nothing is masked after.
-					type={def.secret ? "password" : "text"}
-					placeholder={def.placeholder}
-					spellCheck={false}
-					autoComplete="off"
-					value={value}
-					onChange={(e) => onChange(e.target.value)}
-				/>
-				{children}
-			</div>
-		</div>
-	);
-}
-
-// One value the operator has to retype on a phone, with a button that saves the
-// retyping. A blocked clipboard is silent: the text stays selectable.
-function CopyRow({
-	testId,
-	label,
-	value,
-	t,
-}: {
-	testId: string;
-	label: string;
-	value: string;
-	t: TFunction;
-}) {
-	if (value === "") return null;
-	return (
-		<div className="flex items-center gap-2 flex-wrap">
-			<span className="text-xs text-(--text-muted)">{label}</span>
-			<code
-				className="text-xs text-(--text-primary) select-all break-all"
-				style={{ fontFamily: "var(--font-mono)" }}
-			>
-				{value}
-			</code>
-			<CopyButton
-				variant="label"
-				text={value}
-				testId={testId}
-				ariaLabel={`${t("common.copy")}: ${label}`}
-			/>
-		</div>
 	);
 }
